requirements.txt
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
Pillow==9.4.0
PyYAML==6.0
torch==1.13.1
torchaudio==0.13.1
torchvision==0.10.0
tqdm==4.65.0
typing-extensions==4.5.0
